requirements.txt
flask
requests
pandas
google-cloud-bigquery
google-cloud-storage
gunicorn
scikit-learn
xgboost
joblib
matplotlib